### 2023-05-05 #72 @ViniciusLima94
 [Re] Inter-areal balanced amplification enhances signal propagation in a large-scale circuit model of the primate cortex
**Original article:** Joglekar, Madhura R., et al. "Inter-areal balanced amplification enhances signal propagation in a large-scale circuit model of the primate cortex." Neuron 98.1 (2018): 222-234.

**Scientific domain:** Computational Neuroscience
**Programming language:**  Python (NEST SIMULATOR)
**Suggested editor:** @rougier
**Suggested reviewers:** Hans Ekkehard Plesser, Robin Gutzen, Vahid Rostami, Marcel Stimberg
