### 2023-05-28 #72 @benoit-girard
Dear @ViniciusLima94 the provided URLs end-up on "error 404", can you do something about that?
