### 2023-06-05 #72 @benoit-girard
@damiendr and @rgutzen : would you be interested (and available) to review this submission to ReScience C?
