### 2023-06-08 #72 @benoit-girard
@mstimberg : would you be interested (and available) to review this submission to ReScience C?
