### 2023-07-03 #72 @mstimberg
:wave: @ViniciusLima94, @benoit-girard, please find my review below:

I have reviewed the paper/code and can confirm that the code reproduces the figures of the ReScience manuscript. I can also confirm that these results replicate the results from the original paper (within reasonable bounds). In general, this is a well executed replication, the paper is clearly written and convincingly discusses the issues encountered during the replication. The availability of the source code for the original paper remains a bit unclear to me (see below). This and missing instructions in the code repository are the major issues that IMHO should be addressed before publication. I also identified a number of minor issues, see the list below

### Major issues
- **Availability of original source code:** Most of the ReScience manuscript reads as if the Brian 2 source code of the original paper is not available (it does not seem to be linked from the original article), and the authors therefore had to write the NEST implementation from scratch based on the description in the paper (and guessing/adjusting a few parameters that weren't explicitly mentioned). But then I was confused by the comment on p.12: "Furthermore, we noticed the same phenomenon when running the original Brian 2 code for this model (data not shown)." Does that mean that the authors had access to the original Brian 2 code, e.g. by asking the authors for it? If yes, why did this not clear up the questions of the unspecified parameters? In either case, I think the (un)availability of the original source code should be stated near the beginning of the manuscript, possibly mentioning the unreferenced Matlab code that the authors found on GitHub as well.
- **Installation/run instructions**: The code repository currently does not contain any README file or other explication on how to install the necessary requirements and to run the code. Everything is rather straightforward (conda environment, `main.py` file that can run the various protocols, etc.), but there should be step-by-step instructions how to install and run everything. The authors provide two conda environments, one with detailed pinned versions (which I used) and one which only states the general dependencies. I think the latter needs to specify that the NEST versions needs to be at least version 3.0 which integrates the adapted rate model. The authors state that "it is recommended to use this updated version", but I think it is not only "recommended" but necessary to run the code?
  If I am not mistaken, NEST simulations are only reproducible by fixing both the seed and the number of threads used for simulation. This should be made explicit in the instructions as well.

### Minor issues
#### Paper
- p.1: "Here we replicate a recent modelling by" → "a recent modelling **study** by"
- p.1: "One of the key contributions Joglekar et. al. study"  → "**of** the Joglekar et al. study" (also in several places: "et. al." should be "et al.")
- p.2: "with the code related to this article in: XXXXXX" needs a URL
- Figure 1: I'd again state here what "FLN" stands for, since someone might look at the figure before reading the rest of the text. Also, you might consider plotting FLN instead of log(FLN) and use a logarithmic color map, so that the colorbar values would be 10², 10⁴, etc. A bit clearer IMO, but of course there's nothing wrong with the current figure.
- Figure 3f: The legend in the figure seems to have swapped the labels for the dotted vs. solid lines
- References: Readers would appreciate DOIs for the references, I think
#### Code
- Even after looking at the source code, I am not quite what "protocol 3" does. The comment states "compute average frequencies per population in each stage", but there does not seem to be any output/figure at the end of the simulation
- "protocol 4" is an option in `main.py` but does not do anything
- The generated file names 5, 6, 7 correspond to figures 4, 5, 6 in the paper, which is a bit confusing
- One of the generated files has "joglekar" in the name (`fig3_joglekar_e`) while all the others are simply called `fig_3f`, etc.
