### 2023-07-04 #72 @ViniciusLima94
Thank you very much for you review and commends on our manuscript @mstimberg . I'll start to address the points raised as soon as possible, @benoit-girard .
