+++ code/README.md
@@ -1,0 +1,254 @@
+## Introduction
+
+This is a reference implementation of the following model:
+
+Joglekar, Madhura R., et al. "Inter-areal balanced amplification enhances signal propagation in a large-scale circuit model of the primate cortex." Neuron 98.1 (2018): 222-234.
+
+## Platform information
+
+**Platform:** Ubuntu 20.04.2 LTS
+
+**cmake:** 3.16.3
+
+**gcc (GCC):** 9.3.0
+
+**Python:** 3.8.8 
+
+**Matplotlib:** 3.3.4
+
+**NumPy:** 1.18.5
+
+**SciPy:** 1.6.2
+
+**NEST:** 3.0.0
+
+The machine used to run the codes related to this replication was equiped with a processor **Intel Xeon Gold 6320** 2.1GHz with 20C/20T and 125GB ram. The rate models were simulared using a single-core while the spiking-neuron model was simulated using 20C.
+
+The version of the packages for scientific computing in python might be changed without any downside, however for **NEST** the version has to be 3.0 (or above).
+
+## Package installation
+
+To assure that all data and figures will be generated as presented in the article, we recommend following the instructions below and use the modules in the same versions described here, although the code also works in other versions.
+
+### Python installation
+
+The network simulation is implemented with Python (v.3.8.8).
+
+To install Python 3, type in console:
+
+```
+$sudo apt-get update 
+$sudo apt-get install python3.8
+```
+
+### Installing pip
+
+We use pip, a package management system, to install the Python modules described above.
+To install pip in Ubuntu type in a terminal:
+
+```
+$sudo apt install python3-pip
+```
+
+Upgrade pip to the latest version:
+
+```
+$pip3 install --upgrade pip
+```
+
+Installation of packages using pip can be done with the following command:
+
+```
+$pip3 install --user PYTHON_PACKAGE_NAME
+```
+
+### Installation using Anaconda (recommended)
+
+You can install the scientific Python modules with the Anaconda data science platform.
+
+For Linux, Anaconda with Python 3.8 is only available for 64-bit systems. Download link:
+https://repo.anaconda.com/archive/Anaconda3-2020.11-Linux-x86_64.sh
+
+To install open a terminal in the folder containing the downloaded file and type:
+
+```
+$chmod +x Anaconda3-2020.11-Linux-x86_64.sh
+
+$./Anaconda3-2020.11-Linux-x86_64.sh
+```
+
+The easiest way to install the packages required is via the enviroment file that we provided, using the command from the code directory:
+
+```
+conda env create -f rs_joglekar_all.yml
+```
+
+Once the enviroment is installed it can be activate with the command
+
+```
+conda activate rs_joglekar
+```
+
+#### Python modules installation using Anaconda
+
+Alternativelly, the user can manually install the requirements using conda, using the command bellow.
+
+```
+$conda install PYTHON_PACKAGE_NAME
+```
+
+#### Python modules installation using pip
+
+To install the required packages type in terminal:
+
+```
+$pip3 install --user matplotlib==3.3.4
+$pip3 install --user numpy==1.18.5
+$pip3 install --user scipy==1.6.2
+```
+or
+
+```
+$pip3 install --user matplotlib==3.3.4 numpy==1.18.5 scipy==1.6.2
+```
+
+
+### Building NEST version 3
+
+To date the current version of NEST that can be installed via Ananconda is the version 2.20.x, therefore it is necessary to build NEST 3 from the source. To do so first download the tar file with it by using the following link: https://github.com/nest/nest-simulator/archive/refs/tags/v3.0.tar.gz
+
+Next, follow the steps bellow:
+
+#### 1. Decompress the tar file
+```
[…]
