+++ code/README.md
@@ -41,6 +41,43 @@ $sudo apt-get update
 $sudo apt-get install python3.8
 ```
 
+### Installation using Anaconda (**recommended**)
+
+You can install the scientific Python modules with the Anaconda data science platform.
+
+For Linux, Anaconda with Python 3.8 is only available for 64-bit systems. Download link:
+https://repo.anaconda.com/archive/Anaconda3-2020.11-Linux-x86_64.sh
+
+To install open a terminal in the folder containing the downloaded file and type:
+
+```
+$chmod +x Anaconda3-2020.11-Linux-x86_64.sh
+
+$./Anaconda3-2020.11-Linux-x86_64.sh
+```
+
+Once Anaconda is installed, the easiest way to install the packages required is via the enviroment file that we provided, using the command from the code directory:
+
+```
+conda env create -f rs_joglekar_all.yml
+```
+
+Once the enviroment is installed it can be activate with the command
+
+```
+conda activate rs_joglekar
+```
+
+After this step, the codes in this repository can be run withou dependency issues.
+
+#### Python modules installation using Anaconda
+
+Alternativelly, the user can manually install the requirements using conda, using the command bellow.
+
+```
+$conda install PYTHON_PACKAGE_NAME
+```
+
 ### Installing pip
 
 We use pip, a package management system, to install the Python modules described above.
@@ -56,48 +93,14 @@ Upgrade pip to the latest version:
 $pip3 install --upgrade pip
 ```
 
+#### Python modules installation using pip
+
 Installation of packages using pip can be done with the following command:
 
 ```
 $pip3 install --user PYTHON_PACKAGE_NAME
 ```
 
-### Installation using Anaconda (recommended)
-
-You can install the scientific Python modules with the Anaconda data science platform.
-
-For Linux, Anaconda with Python 3.8 is only available for 64-bit systems. Download link:
-https://repo.anaconda.com/archive/Anaconda3-2020.11-Linux-x86_64.sh
-
-To install open a terminal in the folder containing the downloaded file and type:
-
-```
-$chmod +x Anaconda3-2020.11-Linux-x86_64.sh
-
-$./Anaconda3-2020.11-Linux-x86_64.sh
-```
-
-The easiest way to install the packages required is via the enviroment file that we provided, using the command from the code directory:
-
-```
-conda env create -f rs_joglekar_all.yml
-```
-
-Once the enviroment is installed it can be activate with the command
-
-```
-conda activate rs_joglekar
-```
-
-#### Python modules installation using Anaconda
-
-Alternativelly, the user can manually install the requirements using conda, using the command bellow.
-
-```
-$conda install PYTHON_PACKAGE_NAME
-```
-
-#### Python modules installation using pip
 
 To install the required packages type in terminal:
 
@@ -112,10 +115,16 @@ or
 $pip3 install --user matplotlib==3.3.4 numpy==1.18.5 scipy==1.6.2
 ```
 
+### Installing NEST version 3 (or higher)
+
+If the user opts for not using the available enviroment .yml file, NEST can be installed using conda by following the instructions in the NEST official website: https://nest-simulator.readthedocs.io/en/v2.18.0/installation/conda_install.html
+
 
 ### Building NEST version 3
 
-To date the current version of NEST that can be installed via Ananconda is the version 2.20.x, therefore it is necessary to build NEST 3 from the source. To do so first download the tar file with it by using the following link: https://github.com/nest/nest-simulator/archive/refs/tags/v3.0.tar.gz
+Another option is to install NEST from the source.
+
+To do so first download the tar file with it by using the following link: https://github.com/nest/nest-simulator/archive/refs/tags/v3.0.tar.gz
 
 Next, follow the steps bellow:
 
@@ -206,49 +215,9 @@ The outcome of each protocol is detailed below:
 
[…]
