+++ content.tex
@@ -5,12 +5,16 @@
 
 Information processing in the cortex relies on stable and, at the same time, reconfigurable signal propagation among its areas, i.e.: first, the signal is neither attenuated nor amplified as it propagates across the cortical hierarchy of areas; second, the extent to which signal propagate through hierarchically connected regions must be regulated by context. Despite numerous studies done on inter-area signal propagation in the mammalian cortex \cite{van2002fast, kumar2008conditions,kumar2010spiking}, the underlying mechanisms that could allow a simultaneously reliable and flexible signal transmission are not yet completely understood. In computational neuroscience, many models address this question, however most of them make  unrealistic assumptions such as perfect identity of all areas, homogeneity of connection weights across all stages of the hierarchy, and strictly feedforward network architecture,  even though cortical networks are rich in recurrent connections and feedback loops \cite{markov2014b}.
 
-\indent Here we replicate a recent modelling by Joglekar et. al., \cite{joglekar2018inter}. In this work, the authors address the problem of reliable and flexible signal propagation in a large-scale network model embedding a realistic multi-regional connectome. The network structure of the model is prescribed by an empiricaly-determined directed and weighted inter-areal connectivity matrix for the macaque cortex, containing $29$ areas distributed across the occipital, temporal, parietal, and frontal lobes \cite{markov2014b}. Furthermore, regions have heterogeneous time constants, due to a phenomenological reproduction of varying number of excitatory spines across the cortical hierarchy \cite{Elston2007}. The dynamics of network nodes are described initially in terms of a population-level mean-field rate model \cite{chaudhuri2015large,mejias2016feedforward}. In a second part of the article, a spiking network model of leaky integrate-and-fire neurons is used as well to study synchrony modulations within each region.  One of the key contributions Joglekar et. al. \cite{joglekar2018inter} study is to reveal the important role of a mechanism named \emph{global balanced amplification (GBA)}. Briefly, this mechanism consists of increasing inter-areal excitation to facilitate signal propagation to different areas, combined simultaneously with an increase of local inhibition to assure network stability. Modulating GBA provides thus a way to enable stable signal propagation or to disable it, thus conditionally switching on or off the access to regions distant from the site of input injection.
[…]
