+++ bibliography.bib
@@ -27,7 +27,9 @@
   number={6},
   pages={1477--1498},
   year={1990},
-  publisher={American Physiological Society Bethesda, MD}
+  publisher={American Physiological Society Bethesda, MD},
+ doi = {10.1152/jn.1990.63.6.1477 },
+  url = {http://dx.doi.org/10.1152/jn.1990.63.6.1477 }
 }
 
 @techreport{jordan2019nest,
@@ -53,7 +55,9 @@
   journal={Elife},
   volume={8},
   year={2019},
-  publisher={eLife Sciences Publications, Ltd}
+  publisher={eLife Sciences Publications, Ltd},
+ doi = {10.7554/eLife.47314},
+  url = {http://dx.doi.org/10.7554/eLife.47314}
 }
 
 @article{van2002fast,
@@ -64,7 +68,9 @@
   number={5},
   pages={1956--1966},
   year={2002},
-  publisher={Soc Neuroscience}
+  publisher={Soc Neuroscience},
+  doi = {10.1523/JNEUROSCI.22-05-01956.2002},
+  url = {https://doi.org/10.1523/JNEUROSCI.22-05-01956.2002}
 }
 
 @article{kumar2008conditions,
@@ -75,7 +81,9 @@
   number={20},
   pages={5268--5280},
   year={2008},
-  publisher={Soc Neuroscience}
+  publisher={Soc Neuroscience},
+  doi = {10.1523/JNEUROSCI.2542-07.2008},
+  url = {http://dx.doi.org/10.1523/JNEUROSCI.2542-07.2008}
 }
 
 @article{kumar2010spiking,
@@ -86,7 +94,9 @@
   number={9},
   pages={615--627},
   year={2010},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi = {10.1038/nrn2886},
+  url = {http://dx.doi.org/10.1038/nrn2886}
 }
 
 @software{hahne_jan_2021_4739103,
@@ -175,7 +185,9 @@ month = {01}
   number={1},
   pages={17--36},
   year={2014},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi = {10.1093/cercor/bhs270},
+  url = {http://dx.doi.org/ 10.1093/cercor/bhs270 }
 }
 
 @article{joglekar2018inter,
@@ -186,7 +198,9 @@ month = {01}
   number={1},
   pages={222--234},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.neuron.2018.02.031},
+  url = {http://dx.doi.org/10.1016/j.neuron.2018.02.031}
 }
 
 @article{chaudhuri2015large,
@@ -197,7 +211,9 @@ month = {01}
   number={2},
   pages={419--431},
   year={2015},
-  publisher={Elsevier}
+  publisher={Elsevier},
+ doi = {10.1016/j.neuron.2015.09.008},
+  url = {http://dx.doi.org/10.1016/j.neuron.2015.09.008}
 }
 
 @article{mejias2016feedforward,
@@ -206,7 +222,9 @@ month = {01}
   journal={Sci. Adv,},
   volume={2},
   pages={e1601335},
-  year={2016}
+  year={2016},
+ doi = {10.1126/sciadv.1601335},
+  url = {http://dx.doi.org/10.1126/sciadv.1601335}
 }
 
 
@@ -313,7 +331,9 @@ doi = {https://doi.org/10.1016/B0-12-370878-8/00164-6},
 url = {https://www.sciencedirect.com/science/article/pii/B0123708788001646},
 author = {G.N. Elston},
[…]
