### 2023-07-11 #72 @ViniciusLima94
Dear @mstimberg ,

We thank you for reviewing our manuscript, and for the suggestions and corrections that you suggest. As it is usual, changes in the manuscript will be indicated with blue text, and bellow we list your comments followed by our responses.

### Major issues

> **Availability of original source code: Most of the ReScience manuscript reads as if the Brian 2 source code of the original paper is not available (it does not seem to be linked from the original article), and the authors therefore had to write the NEST implementation from scratch based on the description in the paper (and guessing/adjusting a few parameters that weren't explicitly mentioned). But then I was confused by the comment on p.12: "Furthermore, we noticed the same phenomenon when running the original Brian 2 code for this model (data not shown)." Does that mean that the authors had access to the original Brian 2 code, e.g. by asking the authors for it? If yes, why did this not clear up the questions of the unspecified parameters? In either case, I think the (un)availability of the original source code should be stated near the beginning of the manuscript, possibly mentioning the unreferenced Matlab code that the authors found on GitHub as well.**

We thank the reviewer for pointing out our lack of clarity regarding the presence of Brian 2 source code associated with the original publication and its use in replicating the results.

We added a section right after the paper’s Introduction named “Availability of code implementations for the original model”, clarifying this issue:

_In the original publication, the authors do not link any associated code repository with their implementation of the models described. However, we found a repository with Matlab and Python (Brian 2) scripts in the following address:  \href{github.com/xjwanglab/JoglekarEtAl2018_Neuron}{github.com/xjwanglab/JoglekarEtAl2018_Neuron}. Despite the fact that those scripts helped to detect discrepancies and non-specified parameters in the original article, for the spiking model, not all parameters could be directly ``plugged-in'' into Nest without conversion, and some adjustment was often required in order to obtain the firing rate regimes specified in the original paper, as we elaborated throughout this replication paper._

In the results section we modified the following two descriptions to make it clear how exactly the authors’ implementation helped us to replicate the model.

Paragraph after equation 27:
_This condition was not mentioned in the original article. However it was found in the Matlab implementation of the model that we cited in the section ``Availability of code implementations for the original model''._

Paragraph after equation 28:
_In the original paper, the authors mention that a background input was given to the network in order to maintain the spiking activity of the excitatory population between $0.75$ and $1.5$ Hz and the inhibitory population between $5$ and$6$ Hz. Although the authors did not specify the background noise parameters, the model equations found in their Brian2 code take the form $dV/dt = f(V) + (\sigma/\tau)\xi(t)$ (``rishimodelpython_brian2_spiking.py''), lines 137 and 141). We converted the values of $\sigma$ set on their code using Eq.ref{eq:noise_input}. However, some adjustments of the parameters were required to reproduce the specified steady-state rates (Table~ref{tab:rates})._

Text added in section 3.3, subsection “Synchronous regime”:
_Furthermore, we noticed the same phenomenon  when running the original Brian 2 code for this model (data not shown), which suggests this is not an issue of incorrectly fitting parameters but rather an intrinsic property of the model. Additionally, we noticed that in the authors' script the seed was fixed (``rishimodelpython_brian2_spiking.py'', lines 30-32)._

> Installation/run instructions: The code repository currently does not contain any README file or other explanation on how to install the necessary requirements and to run the code. Everything is rather straightforward (conda environment, main.py file that can run the various protocols, etc.), but there should be step-by-step instructions on how to install and run everything. The authors provide two conda environments, one with detailed pinned versions (which I used) and one which only states the general dependencies. I think the latter needs to specify that the NEST version needs to be at least version 3.0 which integrates the adapted rate model. The authors state that "it is recommended to use this updated version", but I think it is not only "recommended" but necessary to run the code?
    If I am not mistaken, NEST simulations are only reproducible by fixing both the seed and the number of threads used for simulation. This should be made explicit in the instructions as well.

We thank the reviewer for pointing out our mistake. We now added a detailed README file to the code directory.

### Minor issues

#### Paper

> p.1: "Here we replicate a recent modelling by" → "a recent modelling study by"

We thank the reviewer for pointing out our mistake. The sentence is now corrected.

> p.1: "One of the key contributions Joglekar et. al. study" → "of the Joglekar et al. study" (also in several places: "et. al." should be "et al.")

We thank the reviewer for pointing out our mistake. The sentence is now corrected.

> p.2: "with the code related to this article in: XXXXXX" needs a URL

We thank the reviewer for pointing out our mistake. We have now inserted the link to the following repository: github.com/ViniciusLima94/ReScience-Joglekar/blob/master/code/interareal/

> Figure 1: I'd again state here what "FLN" stands for, since someone might look at the figure before reading the rest of the text. Also, you might consider plotting FLN instead of log(FLN) and use a logarithmic color map, so that the colorbar values would be 10², 10⁴, etc. A bit clearer IMO, but of course there's nothing wrong with the current figure.

We thank the reviewer for the suggestion, we have updated the Figure 1(a), by showing FLN rather than log(FLN), and adopted a logarithmic scale for the colorbar. Additionally, we replaced the title “log(FLN)” in this panel by “Connection strength”.

> Figure 3f: The legend in the figure seems to have swapped the labels for the dotted vs. solid lines

We thank the reviewer for directing our attention to this mistake. We have now corrected it.

> References: Readers would appreciate DOIs for the references, I think

We added the DOIs to the references in the bibtex file however they still do not appear in the paper.

#### Code

> Even after looking at the source code, I am not quite sure what "protocol 3" does. The comment states "compute average frequencies per population in each stage", but there does not seem to be any output/figure at the end of the simulation

We thank the reviewer for directing our attention to this mistake. Indeed, there is no output figure for this protocol since it serves to compute the neurons’ population average firing rate to verify if they lie in the ranges specified by the original paper, as we mention in Section 3.4:

“were the mean firing rate ranges in the absence of input which are: 0.75 to 1.5 Hz for the excitatory population; and 5 to 6 Hz for the inhibitory population.”

The outcomes of this protocol are shown in Table 4.

Additionally, we added the following print commands at the end of protocol 3 (lines 210-218 on main.py):

print(f" Weak GBA ")
print(f" | Async | Sync |")
print(f" E | {rewa}| {rews}|")
print(f" I | {riwa}| {riws}|")
print("------------------------------")
print(f" Strong GBA ")
print(f" | Async | Sync |")
print(f" E | {resa}| {ress}|")
print(f" I | {risa}| {riss}|")

Which will give the average firing rate as values printed in the command line.

> "protocol 4" is an option in main.py but does not do anything

We thank the reviewer for directing our attention to this mistake. We removed protocol 4 from the argparse options.

Before we had protocol four to simulate the Synchronous regime with strong GBA in a case where the activity of the neurons diverge.

Before submission, we moved this analysis to protocol 3, and it is corresponding to the last simulation in this protocol (lines 163-164 in main.py)

i_ss, _, t_ss, _, maxf_ss,_,_ = fig5_6.simulate(lnt = lnt, seed = 10, simtime = simtime, reg = 'sync', gba = 'strong-gba', transient = 0, dt = dt) plot_figures.plot_raster(t_ss, i_ss, None, simtime, 'figures/fig7.png', save=True)

>  The generated file names 5, 6, 7 correspond to figures 4, 5, 6 in the paper, which is a bit confusing

We thank the reviewer for directing our attention to this mistake. We have renamed the corresponding figures to match their order in the paper.

>  One of the generated files has "joglekar" in the name (fig3_joglekar_e) while all the others are simply called fig_3f, etc.

We thank the reviewer for directing our attention to this mistake. We have corrected it.
