### 2023-07-21 #72 @mstimberg
@benoit-girard, @ViniciusLima94
I thank the authors for the changes, almost all of my concerns have been addressed. I was able to run all the code and to reproduce all the figures (with one minor exception, see below) and table 4 following the (very detailed!) instructions in the README file. I did all this based on the conda environment, so I did not review the other installation instructions (NEST from source, pip installs, etc.) in detail, but everything seemed very reasonable from a cursory look.
My figure 5 looks slightly different from the one in the paper, though – could this be an issue with seed and/or number of threads? I ran it with NEST 3.4 and 20 threads (`python main.py 2 20`):

![fig5](https://github.com/ReScience/submissions/assets/1381982/0d36545b-418b-4bf5-8345-4e810918ccf1)

A bit of nitpicking: `fig1_b.pdf` and `fig1_c.pdf` appear as Figure 2b and 2c in the paper :)

I am in general convinced with the explanations regarding the parameters that weren't mentioned in the original paper, but could you comment more on the $\tau$ parameter? You state that you could not find values in the code and used 20ms in the weak LBA and 30ms in the strong LBA condition, with the same values for excitatory and inhbitory cells. From a quick look at the source code in rishimodelpython_brian2_spiking.py, however, it seems that they used 20ms for excitatory and 10ms for inhibitory cells?
https://github.com/xjwanglab/JoglekarEtAl2018_Neuron/blob/42c21da0df79611f62b8aa549b2bacd921c79d61/rishimodelpython_brian2_spiking.py#L38-L39C1

> We added the DOIs to the references in the bibtex file however they still do not appear in the paper.

Um, apologies, apparently displaying them is switched off in the LaTeX template (not quite sure why?).
