### 2023-07-24 #72 @stephanmg
Thanks for the submission @ViniciusLima94

@benoit-girard find my comments below, most of the issues have been fixed meanwhile by @ViniciusLima94 though...

# Major issues

The most major issue I faced was the lack of running instructions / README.
This seems already been fixed in a pretty detailed way now. Very good!

# Minor issues
Paper:
- Typos ("wring distance"), need to double check for them throughout manuscript
- Figure 2. Wouldn't mix red and green in a figure (color-blind)
- Would avoid highlighting in Table 2 but state in legend, same for a later Table
- Figure 2. Fonts seems rather small in legend and axis description
- DOIs are now available? In the recent manuscript I still do not see them.

Code
- Remove superfluous / unused code in main.py
- Has the code been linted? It partly looks like, but would be great to have a consistent formatting throughout. Please double check.

@mstimberg my figure looks also slightly different than yours...  I'm not too familiar with NEST simulator, but I agree it could then be an issue of seeds / number of threads perhaps. NEST 3.5 and number of threads 32 is what I used.
