### 2023-07-24 #72 @mstimberg
> @mstim my figure looks also slightly different than yours... I'm not too familiar with NEST simulator, but I agree it could then be an issue of seeds / number of threads perhaps. NEST 3.5 and number of threads 32.

In the new README they state that they ran with 20 threads, which is what I used. So that might explain the differences between our results (but not the difference to the result in the manuscript). But I just realized that it also depends on the compiler that is used (see [NEST docs](https://nest-simulator.readthedocs.io/en/stable/nest_behavior/random_numbers.html)). I guess in my case (Linux, NEST installed via conda) it uses GCC, but if you or the authors are on macOS, it might rather use clang?

PS: @mstimberg, not `@mstim` :)
