+++ code/fig3.py
@@ -4,166 +4,198 @@ import nest.raster_plot
 import sys
 import scipy.io
 import matplotlib.pyplot as plt
-import numpy             as np
+import numpy as np
 
-#########################################################################################
+##############################################################################
 # Reading connectivity data
-#########################################################################################
-data = np.load('interareal/markov2014.npy', allow_pickle=True).item()
+##############################################################################
+data = np.load("interareal/markov2014.npy", allow_pickle=True).item()
 
 # Graph parameters
-Nareas     = 29             # Number of areas
-Npop       = 2             # Number of local populations (one excitatory and one inhibitory)
-Npop_total = Npop * Nareas # Number of total populations
+Nareas = 29  # Number of areas
+Npop = 2  # Number of local populations (one excitatory and one inhibitory)
+Npop_total = Npop * Nareas  # Number of total populations
 # FLN matrix
-#M      = (flnMat > 0).astype(int) # Binarized
-flnMat = data['FLN']
-M      = (flnMat > 0).astype(int) # Binarized
+# M      = (flnMat > 0).astype(int) # Binarized
+flnMat = data["FLN"]
+M = (flnMat > 0).astype(int)  # Binarized
 # Hierarchy values
-h = np.squeeze(data['Hierarchy'].T)
+h = np.squeeze(data["Hierarchy"].T)
 
-def simulate(lnt = 1, simtime = 1000.0, dt = 0.2, params=None, max_cond = True, seed = 0):
-    #########################################################################################
+
+def simulate(lnt=1, simtime=1000.0, dt=0.2, params=None, max_cond=True,
+             seed=0):
+    ##########################################################################
     # Configure NEST kernel and parameters
-    #########################################################################################
+    ##########################################################################
 
     nest.ResetKernel()
-    nest.SetKernelStatus({
-        'resolution': dt,        # Set simulation resolution
-        'print_time': False,     # Enable printing of simulation progress (-> terminal)
-        'local_num_threads': lnt,  # Use two threads to build & simulate the network
-        'rng_seed' : seed
-    })
+    nest.SetKernelStatus(
+        {
+            "resolution": dt,  
+            "print_time": False,  
+            "local_num_threads": lnt, 
+            "rng_seed": seed,
+        }
+    )
 
-    nest.SetDefaults('threshold_lin_rate_ipn',  {'theta': 0.0, 
-                                                 'alpha': 1e6,
-                                                 'mu': 0.0, 
-                                                 'lambda': 1.0,
-                                                 'sigma': 0.0,
-                                                 'rectify_output': max_cond,
-                                                 'linear_summation': True})
+    nest.SetDefaults(
+        "threshold_lin_rate_ipn",
+        {
+            "theta": 0.0,
+            "alpha": 1e6,
+            "mu": 0.0,
+            "lambda": 1.0,
+            "sigma": 0.0,
+            "rectify_output": max_cond,
+            "linear_summation": True,
+        },
+    )
 
-    #########################################################################################
+    ##########################################################################
[…]
