+++ code/fig4_5_6.py
@@ -1,0 +1,330 @@
+# import NEST & NEST rasterplot
+import nest
+import nest.raster_plot
+import sys
+import scipy.io
+import numpy as np
+import setParams
+
+##############################################################################
+# Reading connectivity data
+##############################################################################
+area_names = [
+    "V1",
+    "V2",
+    "V4",
+    "DP",
+    "MT",
+    "8m",
+    "5",
+    "8l",
+    "TEO",
+    "2",
+    "F1",
+    "STPc",
+    "7A",
+    "46d",
+    "10",
+    "9/46v",
+    "9/46d",
+    "F5",
+    "TEpd",
+    "PBr",
+    "7m",
+    "7B",
+    "F2",
+    "STPi",
+    "PROm",
+    "F7",
+    "8B",
+    "STPr",
+    "24c",
+]
+
+data = np.load("interareal/markov2014.npy", allow_pickle=True).item()
+netwParams_hier = data["Hierarchy"]
+flnMat = data["FLN"]
+M = (flnMat > 0).astype(int)
+delayMat = data["Distances"] / 3.5
+
+
+def simulate(
+    lnt=1,
+    seed=100,
+    simtime=1000.0,
+    reg="async",
+    gba="weak-gba",
+    transient=0,
+    dt=0.2,
+    input_to_v1=True,
+    IV1=None,
+    use_default_noise=True,
+    Ie_mean=284.0,
+    Ii_mean=294.0,
+):
+    ##########################################################################
+    # Getting parameters
+    ##########################################################################
+    (
+        Nareas,
+        NE,
+        NI,
+        N,
+        eta,
+        tau_ex,
+        tau_in,
+        d,
+        p,
+        R,
+        sigma,
+        exc_Cm,
+        in_Cm,
+        std_ex,
+        std_in,
+        params,
+        I_ext,
+        inputs,
+    ) = setParams.get_params_spiking_model(dt, reg=reg, gba=gba)
+
+    ##########################################################################
+    # Configure NEST
+    ##########################################################################
+    # Configure kernel
+    nest.ResetKernel()
+    nest.SetKernelStatus(
+        {
+            "resolution": dt,
+            "print_time": True,
+            "local_num_threads": lnt,
+            "rng_seed": seed,
+        }
+    )
+
+    # Excitatory neurons
+    exc_params = {
+        "C_m": exc_Cm,  # Membrane capacity (pF)
+        "E_L": -70.0,  # Resting membrane potential (mV)
+        "I_e": 0.0,  # External input current (pA)
+        "V_reset": -60.0,  # Reset membrane potential after a spike (mV)
+        "V_th": -50.0,  # Spike threshold (mV)
+        "t_ref": 2.0,  # Refractory period (ms)
+        "tau_m": tau_ex,  # Membrane time constant (ms)
+    }
+
+    # Inhibitory neurons
+    in_params = {
+        "C_m": in_Cm,  # Membrane capacity (pF)
+        "E_L": -70.0,  # Resting membrane potential (mV)
+        "I_e": 0.0,  # External input current (pA)
+        "V_reset": -60.0,  # Reset membrane potential after a spike (mV)
+        "V_th": -50.0,  # Spike threshold (mV)
+        "t_ref": 2.0,  # Refractory period (ms)
+        "tau_m": tau_in,  # Membrane time constant (ms)
+    }
+
+    # set defaults for neuron models
+    nest.CopyModel("iaf_psc_delta", "exc_iaf_psc_delta", exc_params)
+    nest.CopyModel("iaf_psc_delta", "in_iaf_psc_delta", in_params)
+
+    ##########################################################################
+    # Create neurons and devices
+    ##########################################################################
+
+    # Creating neuronal populations
+    pop_e = []
+    pop_i = []
+    for pop in range(Nareas):
[…]
