### 2023-07-27 78b2327
update manuscript and tex files
