+++ code/main.py
@@ -222,7 +222,7 @@ if p == 2:
     )
 
     ##########################################################################
-    # Strong GBA / Synchronous (bad seed=0, using lnt=20)
+    # Strong GBA / Synchronous 
     ##########################################################################
     i_ss, _, t_ss, _, maxf_ss, _, _ = fig4_5_6.simulate(
         lnt=lnt,
