+++ code/rs_joglekar.yml
@@ -1,10 +1,0 @@
-name: rs_joglekar_simp
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - matplotlib
-  - scipy
-  - numpy
-  - nest-simulator
-  - ipython
