### 2023-07-27 #72 @ViniciusLima94
Dear @mstimberg, we thank you again for your remarks and commentaries that allowed us to improve our manuscript. We will answer your new batch of suggestions in topics bellow:

> My figure 5 looks slightly different from the one in the paper, though – could this be an issue with seed and/or number of threads? I ran it with NEST 3.4 and 20 threads (python main.py 2 20):

Thank you for bringing this issue for our attention. We were able to reproduce the same result you had above.
While we don’t have a concrete answer to why this happened, there are a few things we can comment on:

- As you mentioned, it could be an issue related to the NEST version and the C compiler it uses.

- Since we originally run the script using the NEST 3 version, and the environment uses NEST 3.4, the RNG generator could have changed.

- We noticed that by slightly decreasing the inhibition component of the drive currents for the synchronous regime network we could retrieve the original result (running with python main.py 2 20). The new version of the code we pushed has the updated parameter (line 133 of setParams.py). The new values were updated in table 2 of the manuscript.

But in any case, we do confirm that it will depend on the number of threads (seeds set) :

![image](https://github.com/ReScience/submissions/assets/17538901/ad3d0af4-dc2a-40f3-b4ad-cf4927300355)

At the same time, the result in Figure 6, portraying the instability of the network in this regime does not show an unstable firing rate anymore. We could find an unstable state again when using the seed 100 (line 227-235 in main.py) and by running the script with four threads: python main.py 2 4. We tested it in different laptops using the conda environment we provided.

In the new version of the manuscript we added a comment (see section Important information needed during replication) that the results, for instance to each extent the stimulus will propagate and whether the firing rate will be stable, depends on the version of NEST and Python as well as the number of threads used. We comment this on the results section for the synchronous spiking neuronal model:

_“Conversely, this dependence on initial conditions can generate instances in which the network is in a less excitable state, as a consequence activity may not reach areas such as PROm. In other cases, activity will reach even area 24c, without divergence of firing rates, if the network is in a slightly more excitable state.”_

Finally, we recall that the instability effect is also observed by running their Brian 2 code for the synchronous and strong GBA regime (see Figure below obtained by running their code), hence this is not solely due to NEST.

![image](https://github.com/ReScience/submissions/assets/17538901/02ee671a-20e5-45f3-99ed-03242d4624cd)

> A bit of nitpicking: fig1_b.pdf and fig1_c.pdf appear as Figure 2b and 2c in the paper :)

True, we have changed the filename. We have also changed the name of the corresponding functions in “plot_figures.py” (fig1b() -> fig2b(); fig1c() -> fig2c()).

> I am in general convinced with the explanations regarding the parameters that weren't mentioned in the original paper, but could you comment more on the parameters? You state that you could not find values in the code and used 20ms in the weak LBA and 30ms in the strong LBA condition, with the same values for excitatory and inhibitory cells. From a quick look at the source code in rishimodelpython_brian2_spiking.py, however, it seems that they used 20ms for excitatory and 10ms for inhibitory cells?

If I’m not mistaken the missing tau I referred to is the one for the local circuit using a rate model (figure 2 in the paper; and table 2). The link above is for the spiking model, for which we used the values you mentioned (lines 62 and 63 in our setParams.py).

### Additional modifications

- To follow the numbering of figures we also rename the scripts fig1.py and fig5_6.py to fig2.py and fig4_5_6.py, respectively (the mention of those scripts in the README file were changed accordingly).
- We updated the values of $\mu_{V,E}$, $\mu_{V,I}$ in Table 5 to the correct ones.
