### 2023-07-27 #72 @ViniciusLima94
Dear @stephanmg, we thank you for your remarks and commentaries that helped us to improve our ReScience manuscript.

### Paper

> Typos ("wring distance"), need to double check for them throughout manuscript

We double checked the manuscript for the mentioned typo, we found it  in the legend of figure 2 and corrected it.

> Figure 2. Wouldn't mix red and green in a figure (color-blind)

We  updated the colors to blue (with dashed lines), and black in the new version of the manuscript.

> Would avoid highlighting in Table 2 but state in legend, same for a later Table

We opted to keep the highlighting since it gives an immediate view of the missing parameters, but we also comment on the legends in the new version of the manuscript.

> Figure 2. Fonts seems rather small in legend and axis description

We increased the font size of labels/legends/titles to 15 points  in the new version of the manuscript.

> DOIs are now available? In the recent manuscript I still do not see them.

We managed to make DOis visible by setting “doi = true”, in line 305 of rescience.cls.

### Code

> Remove superfluous / unused code in main.py

We have removed unnecessary comments and unused code from this and other scripts.

> Has the code been linted? It partly looks like, but would be great to have a consistent formatting throughout. Please double check.

We improved the code quality by running black and manually adjusting each script where needed.
