### 2023-07-28 #72 @stephanmg
LGTM now, also the reproduction issue above looks good.
