### 2023-07-28 #72 @mstimberg
All looks good from my side as well, thanks for the additional explanations here and in the paper. Of course, I can't leave without some final nitpicking – but please go ahead and take these comments into account as you see fit, no need for another round of review from my side:
- In a freshly checked out repo, the `code/figures` directory does not exist and the plotting code will raise an error trying to write to that directory
- I think `fig3_b_d.pdf` should be `fig3_c_d.pdf` :smile:
- I can reproduce Fig 4 and 5 by running `python main.py 2 20` and Fig 6 by running `python main.py 2 4` – but maybe it would be good to separate the two so that one can exactly reproduce the figures from the paper without overwriting the figures from the other call? Another option would be to give an option to call `main.py` without any number of threads, and explicitly use the thread numbers used by the authors for each plot.

All in all, great work, I was happy to review it :tada:
