### 2023-07-28 0ab0024
remove blue highlit paper
