### 2023-07-28 #72 @ViniciusLima94
Dear @mstimberg ,

We thank you for your final remarks.

> In a freshly checked out repo, the code/figures directory does not exist and the plotting code will raise an error trying to write to that directory

We pushed the figures directory.

> I think fig3_b_d.pdf should be fig3_c_d.pdf

We updated the name of the figure to be "fig3_c_d.pdf ".

> I can reproduce Fig 4 and 5 by running python main.py 2 20 and Fig 6 by running python main.py 2 4 – but maybe it would be good to separate the two so that one can exactly reproduce the figures from the paper without overwriting the figures from the other call? Another option would be to give an option to call main.py without any number of threads, and explicitly use the thread numbers used by the authors for each plot.

Now the simulation that generates Figure 6 is in a separated protocol, we added the new protocol to the README file, where we also indicate the number of threads used to exactly reproduce the figure in our manuscript.

We thank you and @stephanmg for the comments that helped us to improve this code+manuscript considerably :)
