### 2023-10-11 #72 @ViniciusLima94
Dear editors,

Since it has been some time since the approval of our manuscript, we would like to kindly ask if there is a prevision of the publication release?

Also, let us know if there is something in which we can be helpful.

Cordially,
The authors
