### 2023-11-16 #72 @rougier
Sorry for loooonnng delay. Zeonodo API has just changed and I need to adapt the published script but for that I need their input on how to do it. Can you ping me here again on Monday to make sure to not have further delay in publishing? In the meantime, you can try to fill as much data as possible in the metadata.tex.
