+++ metadata.tex
@@ -5,7 +5,7 @@
 \def \codeSWH{}
 \def \dataURL{}
 \def \dataDOI{}
-\def \editorNAME{}
+\def \editorNAME{Benoît Girard}
 \def \editorORCID{}
 \def \reviewerINAME{}
 \def \reviewerIORCID{}
@@ -16,11 +16,11 @@
 \def \datePUBLISHED{}
 \def \articleTITLE{[Re] Inter-areal Balanced Amplification Enhances Signal Propagation in a Large-Scale Circuit Model of the Primate Cortex}
 \def \articleTYPE{Replication}
-\def \articleDOMAIN{}
+\def \articleDOMAIN{Computational Neuroscience}
 \def \articleBIBLIOGRAPHY{bibliography.bib}
 \def \articleYEAR{2019}
-\def \reviewURL{}
-\def \articleABSTRACT{}
+\def \reviewURL{https://github.com/ReScience/submissions/issues/72}
+\def \articleABSTRACT{Understanding reliable signal transmission repre-sents a notable challenge for cortical systems, whichdisplay a wide range of weights of feedforward andfeedback connections among heterogeneous areas.We re-examine the question of signal transmissionacross the cortex in a network model based onmesoscopic directed and weighted inter-areal con-nectivity data of the macaque cortex. Our findingsreveal that, in contrast to purely feedforward propa-gation models, the presence of long-range excitatoryfeedback projections could compromise stablesignal propagation. Using population rate modelsas well as a spiking network model, we find thateffective signal propagation can be accomplishedby balanced amplification across cortical areas whileensuring dynamical stability. Moreover, the activa-tion of prefrontal cortex in our model requires theinput strength to exceed a threshold, which isconsistent with the ignition model of conscious pro-cessing. These findings demonstrate our model asan anatomically realistic platform for investigationsof global primate cortex dynamics.}
 \def \replicationCITE{}
 \def \replicationBIB{}
 \def \replicationURL{}
@@ -34,8 +34,8 @@
 \def \articleNUMBER{}
 \def \articleDOI{}
 \def \authorsFULL{Vinicus Lima, Renan Oliveira Shimoura, Nilton Liuji Kamiji, Demian Battaglia, Antonio Carlos Roque da Silva Filho}
-\def \authorsABBRV{}
-\def \authorsSHORT{}
+\def \authorsABBRV{V. Lima, R.O. Shimoura, N.L. Kamiji, D. Battaglia, A.C. Roque}
+\def \authorsSHORT{VL, ROS, NLK, DM, ACR}
 \title{\articleTITLE}
 \date{}
 \author[1,2,\orcid{0000-0001-7115-9041}]{Vinicius Lima}
