### 2023-11-20 #72 @ViniciusLima94
Dear @rougier ,

I'm pinging you as you requested.

I also just pushed a version of the "metadata.tex", with more information inserted. Let me know if there is anything else I can help with.

Best,
Vinicius.
