### 2023-11-22 #72 @rougier
Thanks. Zenodo problem is fixed. Could you fill in the matadata.yaml instead (since the .tex i generated from .yaml). I'll then download you overleaf and work from here.
