+++ metadata.yaml
@@ -5,83 +5,84 @@
 #  - For a successful replication, it shoudl be prefixed with "[Re]"
 #  - For a failed replication, it should be prefixed with "[¬Re]"
 #  - For other article types, no instruction (but please, not too long)
-title: "ReScience (R)evolution"
+title: [Re] Inter-areal Balanced Amplification Enhances Signal Propagation in a Large-Scale Circuit Model of the Primate Cortex
 
 # List of authors with name, orcid number, email and affiliation
 # Affiliation "*" means contact author
 authors:
-  - name: Konrad Hinsen
-    orcid: 0000-0003-0330-9428
-    email: konrad.hinsen@cnrs.fr
-    affiliations: 1,2
+  - name: Vinicius Lima 
+    orcid: 0000-0001-7115-9041
+    email: vinicius.lima.cordeiro@gmail.com 
+    affiliations: 1,2,*
+
+  - name: Renan O. Shimoura 
+    orcid: 0000-0002-6580-5999
+    email: renanshimoura@gmail.com
+    affiliations: 3,2 
+
+  - name: Nilton L. Kamiji
+    orcid: 0000-0001-5006-6612
+    email: nilton.kamiji@gmail.com 
+    affiliations: 3 
+
+  - name: Demian Battaglia
+    orcid: 0000-0003-2021-7920
+    email: demian.battaglia@gmail.com  
+    affiliations: 1,4 
+
+  - name: Antonio C. Roque
+    orcid: 0000-0003-2021-7920
+    email: demian.battaglia@gmail.com  
+    affiliations: 1,4 
     
-  - name: Nicolas P. Rougier
-    orcid: 0000-0002-6972-589X
-    email: Nicolas.Rougier@inria.fr
-    affiliations: 3,4,5,*      # * is for contact author
 
 # List of affiliations with code (corresponding to author affiliations), name
 # and address. You can also use these affiliations to add text such as "Equal
 # contributions" as name (with no address).
 affiliations:
   - code:    1
-    name:    Centre de Biophysique Moléculaire, CNRS UPR4301
-    address: Orléans, France
+    name:    Aix-Marseille University, Institute for Systems Neuroscience (INS UMR 1106, Inserm)
+    address: Marseille, France
     
   - code:    2
-    name:    Synchrotron SOLEIL, Division Expériences
+    name:    These authors contributed equally to this work
     address: Gif sur Yvette, France
     
   - code:    3
-    name:    INRIA Bordeaux Sud-Ouest
-    address: Bordeaux, France
+    name:    Department of Physics, School of Philosophy, Sciences and Letters of Ribeirão Preto (FFCLRP), University of São Paulo,
+    address: Ribeirao Preto, Brazil
     
   - code:    4
-    name:    LaBRI, Université de Bordeaux, Institut Polytechnique de Bordeaux,
-             Centre National de la Recherche Scientifique, UMR 5800
-    address: Talence, France
-    
-  - code:    5
-    name:    Institut des Maladies Neurodégénératives, Université  de Bordeaux,
-             Centre National de la Recherche Scientifique, UMR 5293
-    address: Bordeaux, France
+    name:    University of Strasbourg Institute for Advanced Studies (USIAS)
+    address: Strasbourg, France
 
 
 # List of keywords (adding the programming language might be a good idea)
-keywords: rescience c, rescience x
+keywords: Computational Neuroscience, Python, NEST 
 
 # Code URL and DOI/SWH (url is mandatory for replication, doi after acceptance)
 # You can get a DOI for your code from Zenodo, or an SWH identifier from
 # Software Heritage.
 #   see https://guides.github.com/activities/citable-code/
 code:
[…]
