### 2023-12-04 #72 @rougier
Here is the sanbox version, @ViniciusLima94 can you check everything's ok (then I'll publish the finale version)
https://sandbox.zenodo.org/record/5947
