### 2023-12-04 #72 @ViniciusLima94
Dear @rougier ,

Thank you.

Everything seems to be fine except for the link to the original publication codes, in the section "Availability of code implementations for the original model".

![image](https://github.com/ReScience/submissions/assets/17538901/3f703b46-01a0-413b-a6e2-fc2b9f5e90ae)
