### 2023-12-04 #72 @rougier
Ah, weird. Was it the case on your version ? By the way, software heritage is a better option for savig  software but we can live with this url. I think there's a latex package to break url but I need to find its name.
