### 2023-12-04 #72 @ViniciusLima94
In the overleaf version was fine.

I saw some solutions suggesting to use the hyperref package.

```
\documentclass{article}
\usepackage{hyperref}

\begin{document}

\url{http://www.example.com/this/is/a/very/long/url/that/should/be/broken}

\end{document}
```
