### 2023-12-04 #72 @rougier
Thank. Hyperref is already loaded in the rescience class. But it needed the `xurl` package. By they way, when you need only the url, you can use `\url{some address}` instead of `\href{some address}{some address}`
