### 2023-12-04 #72 @rougier
If the rest of the paper is ok I'll publish it. Waiting for your go.
