### 2023-12-04 #72 @ViniciusLima94
Great thanks @rougier ,

I didn't know about the \url, thanks

The rest of the paper was fine, so for me it is good to go 😁
