### 2023-12-04 #72 @rougier
Here it is: https://zenodo.org/record/10257800.
It'll appear soon the journal website.
