+++ metadata.yaml
@@ -32,8 +32,8 @@ authors:
 
   - name: Antonio C. Roque
     orcid: 0000-0003-2021-7920
-    email: demian.battaglia@gmail.com  
-    affiliations: 1,4 
+    email: antonior@usp.br
+    affiliations: 3 
     
 
 # List of affiliations with code (corresponding to author affiliations), name
