+++ metadata.yaml
@@ -31,7 +31,7 @@ authors:
     affiliations: 1,4 
 
   - name: Antonio C. Roque
-    orcid: 0000-0003-2021-7920
+    orcid: 0000-0003-1260-4840
     email: antonior@usp.br
     affiliations: 3 
     
