### 2023-12-05 #72 @ViniciusLima94
Dear @rougier ,

I just noticed a problem in one of the authors affiliation.

It is my fault since the info was wrong in the "metadata.yaml" (Antonio C. Roque, is acually affiliation 3). I don' know if it is possible to correct it, I already pushed the "metadata.yaml" with the correction.

I'm really sorry about that.

Best,
Vinicius.
