### 2023-12-05 #72 @ViniciusLima94
Hi @rougier ,

Exactly, I have updated his orcid also on the metadata.yaml.

Thank you very much, and sorry for the trouble.

Best.
