### 2023-12-06 #72 @ViniciusLima94
Yes, it is the one I put on the metadata.yaml,

orcid: 0000-0003-1260-4840
