### 2023-12-06 #72 @rougier
Here is the new DOI: https://zenodo.org/records/10276194
